requirements.txt
numpy
Pillow
noise
requests
flask
pytest
flake8
isort
black